Fix part measurement bottom sheet

+++ modules/ui-part-measure/src/main/java/com/ankitsuda/rebound/ui/measure/part/add_sheet/AddPartMeasurementBottomSheetViewModel.kt
@@ -14,51 +14,83 @@
 
 package com.ankitsuda.rebound.ui.measure.part.add_sheet
 
+import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.ankitsuda.navigation.LOG_ID_KEY
+import com.ankitsuda.navigation.PART_ID_KEY
+import com.ankitsuda.navigation.WORKOUT_ID_KEY
 import com.ankitsuda.rebound.domain.entities.BodyPartMeasurementLog
 import com.ankitsuda.rebound.data.repositories.MeasurementsRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
+import timber.log.Timber
 import java.time.LocalDateTime
 import java.util.*
 import javax.inject.Inject
 
 @HiltViewModel
-class AddPartMeasurementBottomSheetViewModel @Inject constructor(private val measurementsRepository: MeasurementsRepository) :
+class AddPartMeasurementBottomSheetViewModel @Inject constructor(
+    private val handle: SavedStateHandle,
+    private val measurementsRepository: MeasurementsRepository
+) :
     ViewModel() {
+    private val logId = handle.get<String?>(LOG_ID_KEY)
+    private val partId = handle.get<String?>(PART_ID_KEY)
+
     private var _log: MutableStateFlow<BodyPartMeasurementLog?> = MutableStateFlow(null)
-    val log = _log
+
+    private var _isUpdate: MutableStateFlow<Boolean> = MutableStateFlow(false)
+    val isUpdate = _isUpdate.asStateFlow()
 
     private var _fieldValue = MutableStateFlow("")
     val fieldValue = _fieldValue
+
+    init {
+        Timber.d("logId null = ${logId == null}")
+        if (logId != null) {
+            viewModelScope.launch {
+                try {
+                    val log = measurementsRepository.getLog(logId)
+                    _log.value = log
+                    _fieldValue.value = _log.value!!.measurement.toString()
+                } catch (e: Exception) {
+                    e.printStackTrace()
+                }
+            }
+        }
+
+        _isUpdate.value = logId != null
+    }
 
     fun setFieldValue(value: String) {
         _fieldValue.value = value
     }
 
-    suspend fun setLogId(logId: String)/*: BodyPartMeasurementLog? */ {
-        viewModelScope.launch {
-            val log = measurementsRepository.getLog(logId)
-            _log.value = log
-//        return log
-            _fieldValue.value = _log.value!!.measurement.toString()
+    fun saveMeasurement() {
+        if (_isUpdate.value) {
+            updateMeasurement()
+        } else {
+            addMeasurementToDb()
         }
     }
 
-    fun addMeasurementToDb(partId: String) {
-        viewModelScope.launch {
-            measurementsRepository.addMeasurementToDb(fieldValue.value.toFloat(), partId)
-            _fieldValue.value = ""
-            _log.value = null
-        }
-    }
-
-    fun updateMeasurement() {
-        if (log.value != null) {
+    private fun addMeasurementToDb() {
+        if (partId != null) {
             viewModelScope.launch {
-                val mLog = log.value!!
+                measurementsRepository.addMeasurementToDb(fieldValue.value.toFloat(), partId)
+                _fieldValue.value = ""
+                _log.value = null
+            }
+        }
+    }
+
+    private fun updateMeasurement() {
+        if (_log.value != null) {
+            viewModelScope.launch {
+                val mLog = _log.value!!
                 mLog.measurement = fieldValue.value.toFloat()
                 mLog.updatedAt = LocalDateTime.now()
                 measurementsRepository.updateMeasurement(
@@ -71,11 +103,13 @@ class AddPartMeasurementBottomSheetViewModel @Inject constructor(private val mea
     }
 
 
-    fun deleteMeasurementFromDb(logId: String) {
-        viewModelScope.launch {
-            measurementsRepository.deleteMeasurementFromDb(logId)
-            _fieldValue.value = ""
-            _log.value = null
+    fun deleteMeasurementFromDb() {
+        if (logId != null) {
+            viewModelScope.launch {
+                measurementsRepository.deleteMeasurementFromDb(logId)
+                _fieldValue.value = ""
+                _log.value = null
+            }
         }
     }
 }
